+++ CHANGELOG.rst
@@ -11,6 +11,9 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- ``dtool.DataSet.overlays`` property
+- ``dtool.DataSet.empty_overlay`` method
+- ``dtool.DataSet.persist_overlay`` method
 - ``dtool info`` CLI command
 - ``dtool.clickutils.info_from_path`` function
 - ``dtool.DataSet.item_path_from_hash`` method